+++ src/main/java/net/thumbtack/school/elections/dto/response/proposalResponses/GetAllProposalsDtoResponse.java
@@ -3,13 +3,13 @@ package net.thumbtack.school.elections.dto.response.proposalResponses;
 import java.util.Map;
 
 public class GetAllProposalsDtoResponse {
-    private Map<String, Integer> proposalList;
+    private Map<String, Double> proposalList;
 
-    public GetAllProposalsDtoResponse(Map<String, Integer> proposalList) {
+    public GetAllProposalsDtoResponse(Map<String, Double> proposalList) {
         this.proposalList = proposalList;
     }
 
-    public Map<String, Integer> getProposalList() {
+    public Map<String, Double> getProposalList() {
         return proposalList;
     }
 }
